Added mouse events. Fixed polygon numbers; they're floats not ints.

+++ src/com/googlecode/maps3/client/PolygonOptions.java
@@ -17,11 +17,11 @@ public class PolygonOptions extends JavaScriptObject
 	}-*/;
 
 	public final native void setFillColor(String value) /*-{ this.fillColor = value; }-*/;
-	public final native void setFillOpacity(int value) /*-{ this.fillOpacity = value; }-*/;
+	public final native void setFillOpacity(double value) /*-{ this.fillOpacity = value; }-*/;
 	
 	public final native void setStrokeColor(String value) /*-{ this.strokeColor = value; }-*/;
-	public final native void setStrokeOpacity(int value) /*-{ this.strokeOpacity = value; }-*/;
-	public final native void setStrokeWeight(int value) /*-{ this.strokeWeight = value; }-*/;
+	public final native void setStrokeOpacity(double value) /*-{ this.strokeOpacity = value; }-*/;
+	public final native void setStrokeWeight(double value) /*-{ this.strokeWeight = value; }-*/;
 
 	/** Need to eventually make an MVCArray version */
 	public final native void setPath(JsArray<LatLng> value) /*-{
